+++ packages/xo-server-audit/src/index.js
@@ -69,6 +69,22 @@ const DEFAULT_BLOCKED_LIST = {
 }
 
 const LAST_ID = 'lastId'
+
+// interface Db {
+//   lastId: string
+//   [RecordId: string]: {
+//     data: object
+//     event: string
+//     id: strings
+//     previousId: string
+//     subject: {
+//       userId: string
+//       userIp: string
+//       userName: string
+//     }
+//     time: number
+//   }
+// }
 class Db extends Storage {
   constructor(db) {
     super()
@@ -100,6 +116,16 @@ class Db extends Storage {
   }
 }
 
+export const configurationSchema = {
+  type: 'object',
+  properties: {
+    active: {
+      description: 'Whether to save user actions in the audit log',
+      type: 'boolean',
+    },
+  },
+}
+
 const NAMESPACE = 'audit'
 class AuditXoPlugin {
   constructor({
@@ -127,6 +153,19 @@ class AuditXoPlugin {
 
     this._auditCore = undefined
     this._storage = undefined
+
+    this._listeners = {
+      'xo:audit': this._handleEvent.bind(this),
+      'xo:postCall': this._handleEvent.bind(this, 'apiCall'),
+    }
+  }
+
+  configure({ active = false }, { loaded }) {
+    this._active = active
+
+    if (loaded) {
+      this._addListeners()
+    }
   }
 
   async load() {
@@ -139,8 +178,7 @@ class AuditXoPlugin {
       this._storage = undefined
     })
 
-    this._addListener('xo:postCall', this._handleEvent.bind(this, 'apiCall'))
-    this._addListener('xo:audit', this._handleEvent.bind(this))
+    this._addListeners()
 
     const exportRecords = this._exportRecords.bind(this)
     exportRecords.permission = 'admin'
@@ -190,34 +228,44 @@ class AuditXoPlugin {
   }
 
   unload() {
+    this._removeListeners()
     this._cleaners.forEach(cleaner => cleaner())
     this._cleaners.length = 0
   }
 
-  _addListener(event, listener_) {
-    const listener = async (...args) => {
-      try {
-        await listener_(...args)
-      } catch (error) {
-        log.error(error)
-      }
+  _addListeners(event, listener_) {
+    this._removeListeners()
+
+    if (this._active) {
+      const listeners = this._listeners
+      Object.keys(listeners).forEach(event => {
+        this._xo.addListener(event, listeners[event])
+      })
     }
-    const xo = this._xo
-    xo.on(event, listener)
-    this._cleaners.push(() => xo.removeListener(event, listener))
   }
 
-  _handleEvent(event, { userId, userIp, userName, ...data }) {
-    if (event !== 'apiCall' || !this._blockedList[data.method]) {
-      return this._auditCore.add(
-        {
-          userId,
-          userIp,
-          userName,
-        },
-        event,
-        data
-      )
+  _removeListeners() {
+    const listeners = this._listeners
+    Object.keys(listeners).forEach(event => {
+      this._xo.removeListener(event, listeners[event])
+    })
+  }
+
+  async _handleEvent(event, { userId, userIp, userName, ...data }) {
+    try {
+      if (event !== 'apiCall' || !this._blockedList[data.method]) {
+        return await this._auditCore.add(
+          {
+            userId,
+            userIp,
+            userName,
+          },
+          event,
+          data
+        )
+      }
+    } catch (error) {
+      log.error(error)
     }
   }
 
@@ -246,7 +294,7 @@ class AuditXoPlugin {
         (req, res) => {
           res.set({
             'content-disposition': 'attachment',
-            'content-type': 'application/json',
+            'content-type': 'application/x-gzip',
           })
           return fromCallback(
             pipeline,
@@ -259,7 +307,7 @@ class AuditXoPlugin {
         {
           suffix: `/audit-records-${new Date()
             .toISOString()
-            .replace(/:/g, '_')}.gz`,
+            .replace(/:/g, '_')}.ndjson.gz`,
         }
       )
       .then($getFrom => ({