+++ packages/xo-server/src/collection.js
@@ -136,7 +136,6 @@ export default class Collection extends EventEmitter {
   }
 
   exists(properties) {
-    /* jshint eqnull: true */
     return this.first(properties).then(model => model !== undefined)
   }