+++ packages/xo-server/src/xo-mixins/remotes.js
@@ -60,6 +60,10 @@ export default class {
       remote = await this._getRemote(remote)
     }
 
+    if (remote.proxy !== undefined) {
+      throw new Error('cannot get handler to proxy remote')
+    }
+
     if (!remote.enabled) {
       throw new Error('remote is disabled')
     }