+++ packages/xo-web/src/xo-app/backup/overview/tab-jobs.js
@@ -117,11 +117,17 @@ const SchedulePreviewBody = decorate([
         let lastRunLog
         for (const runId in logs) {
           const log = logs[runId]
-          if (
-            log.scheduleId === schedule.id &&
-            (lastRunLog === undefined || lastRunLog.start < log.start)
-          ) {
-            lastRunLog = log
+          if (log.scheduleId === schedule.id) {
+            if (log.status === 'pending') {
+              lastRunLog = log
+              break
+            }
+            if (
+              lastRunLog === undefined ||
+              (lastRunLog.end || lastRunLog.start) < (log.end || log.start)
+            ) {
+              lastRunLog = log
+            }
           }
         }
         cb(lastRunLog)
@@ -342,7 +348,8 @@ class JobsTable extends React.Component {
         icon: 'preview',
       },
       {
-        handler: (job, { goToNewTab }) => goToNewTab(`/backup/${job.id}/edit`),
+        handler: (job, { goTo, goToNewTab, main }) =>
+          (main ? goTo : goToNewTab)(`/backup/${job.id}/edit`),
         label: _('formEdit'),
         icon: 'edit',
         level: 'primary',